+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## 24.11.1
 * Added content configuration interface that has `setGlobalContentCallback` to get notified about content changes.
 
+* Updated underlying Android SDK version to 24.7.6
+* Updated underlying iOS SDK version to 24.7.8
+
 ## 24.11.0
 * Added further intent redirection vulnerability checks in Android
 * Added support for Android 15 (API level 35)